+++ CssUI/CSS/Tokens/Complex/NumberToken.cs
@@ -25,7 +25,7 @@ public sealed class NumberToken : ValuedTokenBase
     /// <summary>
     /// Gets the value as an integer. Only valid when <see cref="DataType"/> is <see cref="ENumericTokenType.Integer"/>.
     /// </summary>
-    public int AsInteger => data.IntegerValue;
+    public long AsInteger => data.IntegerValue;
 
     /// <summary>
     /// Gets the value as a double. Works for both integer and number types.
@@ -34,7 +34,7 @@ public sealed class NumberToken : ValuedTokenBase
         ? data.IntegerValue
         : data.NumberValue;
 
-    public NumberToken(ENumericTokenType DataType, ReadOnlySpan<char> Value, int number) : base(ECssTokenType.Number, Value)
+    public NumberToken(ENumericTokenType DataType, ReadOnlySpan<char> Value, long number) : base(ECssTokenType.Number, Value)
     {
         this.DataType = DataType;
         this.data = NumericTokenData.FromInteger(number);
@@ -54,12 +54,11 @@ public sealed class NumberToken : ValuedTokenBase
         this.DataType = DataType;
         if (DataType == ENumericTokenType.Integer)
         {
-            // Handle both int and long from tokenizer's ToInteger which returns long
-            int intValue = number switch
+            long intValue = number switch
             {
                 int i => i,
-                long l => (int)l,
-                _ => Convert.ToInt32(number)
+                long l => l,
+                _ => Convert.ToInt64(number)
             };
             this.data = NumericTokenData.FromInteger(intValue);
         }
